the counter can be a u16 with no practical problems

+++ beeps_core/src/hlc.rs
@@ -17,8 +17,8 @@ pub struct Hlc {
 
     /// A counter we increment if the new physical time is equal to or less than
     /// the previous. Second to break ties.
-    #[cfg_attr(test, proptest(strategy = "0..=10u64"))]
-    counter: u64,
+    #[cfg_attr(test, proptest(strategy = "0..=10u16"))]
+    counter: u16,
 
     /// The node ID of the replica that generated this HLC. Third to break ties.
     node: NodeId,
@@ -139,7 +139,7 @@ impl Hlc {
     }
 
     /// Get the counter of this HLC.
-    pub fn counter(&self) -> u64 {
+    pub fn counter(&self) -> u16 {
         self.counter
     }
